Highly rework rainbow code and add lots of functionality
Now every figure may be drawn using 11 rainbow types (9 real, one NONE
and one BEST).

There are also 3 concentrical rainbow subtypes, for rectangles, rhombs
and circles.

Grid now has a central Rainbow object to draw with and methods to set and
unset it.

The rainbow color manipulation is done on Grid::set_color that transforms
the given color according to the defined Rainbow object. Added support to
ignore rainbow in Grid::set_color, that is useful to draw background.

Currently the circular rainbow type is close, but not equal to the circle
drawing algorithm. So keep the old draw_*_rainbow methods for reference
until we synchronize both. Then the draw_*_rainbow methods could be
removed as redundant, since this can be achieved by other methods
(usually draw_filled_*) and RAINBOW_BEST type.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>