Rework undo and redo, do not add duplicate layers
By default (this is what GUI and most users want) start new undo layer
only on the first change and not unconditionally on push_undo.

However strict test applications may want to support the mode when redo
layers are removed on push_undo, and when undo is done only to the last
requested point (push_undo), support this by adding boolean argument to
undo and push_undo.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>