Fix redo and undo to work correctly for both tests and GUI
The requirements of GUI is not to duplicate undo layers. The requirement
of tests is to always create new undo layer even if nothing was changed.
To satisfy both, implement the proper solution to force undo layer
immediately when specially requested, and do it on the change by default.

Fix unintentionally broken tests with manual-last option.

Fix unintentionally broken push_undo.

Now everything seems to work as expected for both tests and GUI. Except
for minor case with non-atomic draw_filled_<figure>_2 methods when the
outline changes color temporarily and then restores it back. When calling
twice, it fills the undo layer altought the end result was not changed.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>