Fix Grid::move and add more types (MOVE, COPY, SWAP)
The 4-th type (STEP) is currently undefined, not implemented.

Do not erase with bg_color when not needed (erase only the source area).

Copy does not erase with bg_color at all, leaves the source area as is.

Swap replaces source and destination areas.

Add unit test and manual test for move method.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>